--- app/src/main/java/com/projectlily/wonderreader/services/BluetoothService.kt
@@ -1,13 +1,0 @@
-package com.projectlily.wonderreader.services
-
-import com.projectlily.wonderreader.types.JsonEvent
-
-class BluetoothService {
-    fun send(jsonEvent: JsonEvent) {
-        return
-    }
-
-    fun onReceive() {
-        return;
-    }
-}
